moved QNatspec and TransactDialog to alethzero, reduced size of libaleth by 3 times
