+++ alethzero/AccountHolder.cpp
@@ -25,7 +25,7 @@
 #include <libdevcore/Log.h>
 #include <libethcore/KeyManager.h>
 #include <libwebthree/WebThree.h>
-#include <libaleth/QNatspec.h>
+#include "QNatspec.h"
 #include <libaleth/AlethFace.h>
 #include "ZeroFace.h"
 using namespace std;
