+++ alethzero/QNatspec.cpp
