+++ libaleth/CMakeLists.txt
@@ -9,7 +9,7 @@ file(GLOB UI_RESOURCES "*.ui")
 set(EXECUTABLE aleth)
 add_library(${EXECUTABLE} ${SRC_LIST} ${HEADERS} ${UI_RESOURCES})
 set_target_properties(${EXECUTABLE} PROPERTIES AUTOUIC ON AUTOMOC ON)
-eth_use(${EXECUTABLE} REQUIRED Qt::Core Qt::Widgets Qt::WebEngine Web3::webthree Eth::devcore Eth::p2p Eth::devcrypto Eth::ethereum Eth::natspec)
+eth_use(${EXECUTABLE} REQUIRED Qt::Core Qt::Widgets Web3::webthree Eth::devcore Eth::p2p Eth::devcrypto Eth::ethereum Eth::natspec)
 
 install( TARGETS ${EXECUTABLE} RUNTIME DESTINATION bin ARCHIVE DESTINATION lib LIBRARY DESTINATION lib )
 install( FILES ${HEADERS} DESTINATION include/${EXECUTABLE} )
